Remove special "is_in_square_brackets" behavior
This removes the behavior when matching a sequence of items
enclosed in square brackets. Prior to this change, the items
would be made independently optional. For example, [-a -o]
would allow -a OR -o to be specified. With this change, the
square brackets make the entire sequence optional, not individual
items. So if -a is specified, -o must be as well.