Correct handling of -- to allow it to be an option-argument
Prior to this change, an argv like `--foo -- --` would treat
the first -- as signalling the end of options (all remaining args
are positional). Now we correctly allow the first -- to be the
option-argument, as OpenGroup demands.

This required factoring the output of argument classification
into a struct, to give us a place to hang the index of the
-- that signals the end of options. (Previously we just assumed
it was the first --).